+++ CHANGES.md
@@ -14,6 +14,8 @@ Please follow the [`CHANGES.md` modification guidelines](https://github.com/dcos
 
 ### Notable changes
 
+* Upgraded to CockroachDB 19.1 (D2IQ-69872)
+
 
 ### Fixed and improved